fix: use OrganizationSlugType for org slugs
