+++ libs/portals/core/src/types/portalCore.ts
@@ -7,6 +7,7 @@ import { RouteObject } from 'react-router-dom'
 import type { Features } from '@island.is/react/feature-flags'
 import { IconProps } from '@island.is/island-ui/core'
 import { User } from '@island.is/shared/types'
+import { OrganizationSlugType } from '@island.is/shared/constants'
 
 /**
  * A navigational item used by the service portal
@@ -61,7 +62,7 @@ export interface PortalNavigationItem {
   /**
    * Service provider id used for displaying service provider information in the module.
    */
-  serviceProvider?: string
+  serviceProvider?: OrganizationSlugType
 }
 
 /**
