+++ libs/service-portal/core/src/utils/constants.ts
@@ -1,32 +1,4 @@
-export type ProviderSlugType =
-  | 'haskoli-islands'
-  | 'vinnueftirlitid'
-  | 'thjodskra'
-  | 'stafraent-island'
-  | 'samgongustofa'
-  | 'husnaedis-og-mannvirkjastofnun'
-  | 'fjarsysla-rikisins'
-  | 'stjornarrad-islands'
-  | 'rikislogreglustjori'
-  | 'personuvernd'
-  | 'tryggingastofnun'
-  | 'sjukratryggingar'
-  | 'thjodskra-islands'
-  | 'fiskistofa'
-  | 'domstolasyslan'
-  | 'samband-islenskra-sveitafelaga'
-  | 'skatturinn'
-  | 'syslumenn'
-  | 'domsmalaraduneytid'
-  | 'utlendingastofnun'
-  | 'menntamalastofnun'
-  | 'umhverfisstofnun'
-  | 'sjukratryggingar-islands'
-  | 'vegagerdin'
-  | 'landlaeknir'
-  | '/s/islandis'
-
-export const ICELAND_SLUG = 'syslumenn'
+export const SYSLUMENN_SLUG = 'syslumenn'
 export const ISLANDIS_SLUG = 'stafraent-island'
 export const MENNTAMALASTOFNUN_SLUG = 'menntamalastofnun'
 export const FJARSYSLAN_SLUG = 'fjarsysla-rikisins'
@@ -37,7 +9,7 @@ export const RIKISLOGREGLUSTJORI_SLUG = 'rikislogreglustjori'
 export const UMHVERFISSTOFNUN_SLUG = 'umhverfisstofnun'
 export const VINNUEFTIRLITID_SLUG = 'vinnueftirlitid'
 export const TRYGGINGASTOFNUN_SLUG = 'tryggingastofnun'
-export const SJUKRATRYGGINGAR_SLUG = 'sjukratryggingar-islands'
+export const SJUKRATRYGGINGAR_SLUG = 'sjukratryggingar'
 export const SKATTURINN_SLUG = 'skatturinn'
 export const APPLICATION_SERVICE_PROVIDER_SLUG = 'stafraent-island'
 export const VEGAGERDIN_SLUG = 'vegagerdin'
