+++ libs/service-portal/education-license/src/screens/EducationLicense/EducationLicense.tsx
@@ -2,7 +2,7 @@ import { defineMessage } from 'react-intl'
 
 import { Box } from '@island.is/island-ui/core'
 import { useLocale, useNamespaces } from '@island.is/localization'
-import { FootNote, ICELAND_SLUG, m } from '@island.is/service-portal/core'
+import { FootNote, SYSLUMENN_SLUG, m } from '@island.is/service-portal/core'
 import { IntroHeader } from '@island.is/service-portal/core'
 import { LicenseCards } from './components/LicenseCards'
 
@@ -19,11 +19,11 @@ function EducationLicense() {
           defaultMessage:
             'Hér er markmiðið að þú getir fundið yfirlit yfir leyfisbréf og vottorð til starfsréttinda. Unnið er að því að koma öllum leyfisbréfum og vottorðum um starfsréttindi á einn stað. Núna birtast leyfisbréf kennara sem hafa verið útskrifaðir frá 1988 sem sótt eru til Menntamálastofnunar.',
         })}
-        serviceProviderSlug={ICELAND_SLUG}
+        serviceProviderSlug={SYSLUMENN_SLUG}
         serviceProviderTooltip={formatMessage(m.occupationalLicenseTooltip)}
       />
       <LicenseCards />
-      <FootNote serviceProviderSlug={ICELAND_SLUG} />
+      <FootNote serviceProviderSlug={SYSLUMENN_SLUG} />
     </Box>
   )
 }
