+++ libs/service-portal/graphql/src/index.ts
@@ -30,5 +30,6 @@ export * from './hooks/profile/useDeleteIslykillValue'
 export * from './hooks/passport/usePassport'
 export * from './hooks/passport/useChildrenPassport'
 export * from './hooks/alertBanners/useAlertBanners'
+export * from './hooks/organizations/useOrganization'
 export * from './hooks/organizations/useOrganizations'
 export * from './schema'
