+++ libs/shared/constants/src/lib/organizationSlug.ts
@@ -25,3 +25,7 @@ export type OrganizationSlugType =
   | 'syslumenn'
   | 'domsmalaraduneytid'
   | 'utlendingastofnun'
+  | 'menntamalastofnun'
+  | 'umhverfisstofnun'
+  | 'vegagerdin'
+  | 'landlaeknir'
